+++ examples/adult/demo/main_test.go
@@ -2,6 +2,9 @@ package main
 
 import (
 	"flag"
+	"github.com/gomlx/gomlx/graph"
+	"github.com/stretchr/testify/require"
+	"os"
 	"testing"
 )
 
@@ -12,6 +15,9 @@ func TestMainFunc(t *testing.T) {
 	}
 	flag.Parse()
 	*flagNumSteps = 10
-	*flagPlatform = "Host"
+	if _, found := os.LookupEnv(graph.DefaultPlatformEnv); !found { // "GOMLX_PLATFORM"
+		err := os.Setenv("GOMLX_PLATFORM", "Host") // Force default to CPU.
+		require.NoError(t, err)
+	}
 	main()
 }